Project structure corrected
