+++ create-post.html
@@ -97,6 +97,6 @@
     </div>
   </footer>
 </div>
-<script type="module" src="js/postCreator.js"></script>
+<script type="module" src="js/createPost.js"></script>
 </body>
 </html>
